+++ docs/api/changelog.rst
@@ -40,6 +40,9 @@ Version 0.10
 * Fix (harmless) traceback in logs which could occur when using
   :meth:`aioxmpp.muc.Room.send_message_tracked`.
 
+* Fix :func:`aioxmpp.service.is_depsignal_handler` and
+  :func:`~aioxmpp.service.is_attrsignal_handler` when used with ``defer=True``.
+
 .. _api-changelog-0.9:
 
 Version 0.9